apps/monitor/requirements.txt
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1
